refactor: auth 관련 클래스 정리
