+++ src/main/java/roomescape/infrastructure/authentication/AuthenticatedProfile.java
@@ -1,4 +1,4 @@
-package roomescape.service.auth;
+package roomescape.infrastructure.authentication;
 
 public record AuthenticatedProfile(String name) {
 }
